Refactored to be modular and use blueprints.
