+++ .flaskenv-example
@@ -4,3 +4,5 @@ FLASK_DEBUG=1
 FLASK_RUN_HOST="localhost"
 FLASK_RUN_PORT="5000"
 SESSION_COOKIE_SECURE=False
+FLASK_ENV=development
+SQLALCHEMY_ECHO=False
