--- templates/history.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-4bw+/aepP/YC94hEpVNVgiZdgIC5+VKNBQNGCHeKRQN+PtmoHDEXuppvnDJzQIu9" crossorigin="anonymous">
-    <link href="{{ url_for('static', filename='style.css') }}" rel="stylesheet">
-    <title>Login</title>
-  </head>
-  <body>
-    <div class="container mt-5">
-      <h1>Conversation History</h1>
-      <div class="card mt-3">
-        <div class="card-body">
-          <h5 class="card-title">Previous Conversations</h5>
-          <!-- Here we loop through each conversation history item -->
-          <ul class="list-group">
-            {% for entry in history %}
-              <li class="list-group-item">
-                <strong>{{ entry.timestamp.strftime('%Y-%m-%d %H:%M:%S') }}</strong>
-                <p>{{ entry.title | safe }}</p>
-              </li>
-            {% else %}
-              <li class="list-group-item">No conversation history found.</li>
-            {% endfor %}
-          </ul>
-        </div>
-    </div>
-    <!-- Optionally, add a link to go back to the main page or other navigation -->
-    <a href="{{ url_for('index') }}" class="btn btn-primary mt-3">Back to Home</a>
-  </div>
-  </body>
-</html>
