--- templates/login.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Montserrat">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-4bw+/aepP/YC94hEpVNVgiZdgIC5+VKNBQNGCHeKRQN+PtmoHDEXuppvnDJzQIu9" crossorigin="anonymous">
-    <link href="{{ url_for('static', filename='style.css') }}" rel="stylesheet">
-    <title>Login</title>
-  </head>
-  <body>
-    <section class="vh-100 bg-dark">
-      <div class="container py-5 h-100">
-        <div class="row d-flex justify-content-center align-items-center h-100">
-          <div class="col col-xl-10">
-            <div class="card" style="border-radius: 1rem;">
-              <div class="row g-0">
-                <div class="col-md-6 col-lg-5 d-none d-md-block">
-                  <img src="{{ url_for('static', filename='images/snowgoose1.png') }}"
-                    alt="login form" class="img-fluid" style="border-radius: 1rem 0 0 1rem;" />
-                </div>
-                <div class="col-md-6 col-lg-7 d-flex align-items-center">
-                  <div class="card-body p-4 p-lg-5 text-black">
-    
-                    <form action="#" method="post">
-    
-                      <div class="d-flex align-items-center mb-3 pb-1">
-                        <img src="{{ url_for('static', filename='images/snowgoose-logo-horizantal-black-100h.png') }}">
-                      </div>
-    
-                      <div class="form-outline mb-4">
-                        <input type="text" name="username" class="form-control form-control-lg" />
-                        <label class="form-label" for="password">Username</label>
-                      </div>
-    
-                      <div class="form-outline mb-4">
-                        <input type="password" name="password" class="form-control form-control-lg" />
-                        <label class="form-label" for="password">Password</label>
-                      </div>
-                      <div class="pt-1 mb-4">
-                        <button class="btn btn-primary" type="submit">Submit</button>
-                      </div>
-                    </form>
-    
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </section>    
-
-
-
-
-  </body>
-</html>
