--- templates/settings.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-4bw+/aepP/YC94hEpVNVgiZdgIC5+VKNBQNGCHeKRQN+PtmoHDEXuppvnDJzQIu9" crossorigin="anonymous">
-    <link href="{{ url_for('static', filename='style.css') }}" rel="stylesheet">
-    <title>Login</title>
-  </head>
-  <body>
-    <script type="text/javascript">
-        function confirmDelete() {
-            return confirm("Are you sure you want to delete this user?");
-        }
-    </script>
-    <div class="container text-center">
-        <div class="row justify-content-center align-items-center">
-            <div class="col-8">
-                <h2>Add User</h2>
-                <form method="post" action="{{ url_for('add_user') }}">
-                    <input type="text" name="username" placeholder="Username">
-                    <input type="password" name="password" placeholder="Password">
-                    <input type="email" name="email" placeholder="Email">
-                    <input type="checkbox" name="is_admin"> Admin?
-                    <button type="submit">Add User</button>
-                </form>
-                <table class="table table-bordered table-striped">
-                    <thead class="thead-dark">
-                        <tr>
-                            <th>ID</th>
-                            <th>Username</th>
-                            <th>Email</th>
-                            <th>Is Admin</th>
-                            <th>Actions</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <!-- Use Jinja2 templating to loop through the users from the context -->
-                        {% for user in users %}
-                        <tr>
-                            <td>{{ user.id }}</td>
-                            <td>{{ user.username }}</td>
-                            <td>{{ user.email }}</td>
-                            <td>{{ 'Yes' if user.is_admin else 'No' }}</td>
-                            <td>
-                                <form method="post" action="{{ url_for('delete_user', id=user.id) }}" onsubmit="return confirmDelete()">
-                                    <button type="submit">Delete User</button>
-                                </form>
-                            </td>
-                        </tr>
-                        {% endfor %}
-                    </tbody>
-                </table>
-            </div>
-        </div>
-    </div>
-  </body>
-</html>
